# web/static/swagger/swagger.yaml
basePath: /api
definitions:
  github_com_alchemillahq_sylve_internal.APIResponse-any:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_alchemillahq_sylve_internal.APIResponse-github_com_alchemillahq_sylve_internal_db_models_BasicSettings:
    properties:
      data:
        $ref: '#/definitions/github_com_alchemillahq_sylve_internal_db_models.BasicSettings'
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_alchemillahq_sylve_internal_db_models.AvailableService:
    enum:
    - dhcp-server
    - jails
    - samba-server
    - virtualization
    - wol-server
    type: string
    x-enum-varnames:
    - DHCPServer
    - Jails
    - SambaServer
    - Virtualization
    - WoLServer
  github_com_alchemillahq_sylve_internal_db_models.BasicSettings:
    properties:
      id:
        type: integer
      initialized:
        type: boolean
      pools:
        items:
          type: string
        type: array
      restarted:
        type: boolean
      services:
        items:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_db_models.AvailableService'
        type: array
    type: object
...
  github_com_alchemillahq_sylve_internal_interfaces_services_jail.AddJailNetworkRequest:
    properties:
      ctId:
        type: integer
      defaultGateway:
        type: boolean
      dhcp:
        type: boolean
      ip4:
        type: integer
      ip4gw:
        type: integer
      ip6:
        type: integer
      ip6gw:
        type: integer
      macId:
        type: integer
      name:
        type: string
      slaac:
        type: boolean
      switchName:
        type: string
    required:
    - ctId
    - name
    - switchName
    type: object
...
  github_com_alchemillahq_sylve_internal_interfaces_services_system.InitializeRequest:
    properties:
      pools:
        items:
          type: string
        type: array
      services:
        items:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_db_models.AvailableService'
        type: array
    required:
    - pools
    - services
    type: object
...
  internal_handlers_jail.SetInheritanceRequest:
    properties:
      ipv4:
        type: boolean
      ipv6:
        type: boolean
    type: object
...
host: sylve.lan:8181
info:
  contact:
    email: hello@alchemilla.io
    name: Alchemilla Ventures Pvt. Ltd.
    url: https://alchemilla.io
  description: Sylve is a lightweight GUI for managing Bhyve, Jails, ZFS, networking,
    and more on FreeBSD.
  license:
    name: BSD-2-Clause
    url: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  termsOfService: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  title: Sylve API
  version: 0.0.1
paths:
...
  /basic/initialize:
    post:
      consumes:
      - application/json
      description: Initialize Sylve with the provided configuration
      parameters:
      - description: Initialization Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_system.InitializeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Initialize Sylve
      tags:
      - Health
  /basic/reboot:
    post:
      consumes:
      - application/json
      description: Initiate a system reboot
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Initiate System Reboot
      tags:
      - Health
  /basic/settings:
    get:
      consumes:
      - application/json
      description: Retrieve the basic settings of Sylve
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-github_com_alchemillahq_sylve_internal_db_models_BasicSettings'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Basic Settings
      tags:
      - Health
...
  /jail/network:
    post:
      consumes:
      - application/json
      description: Add a network switch to a jail
      parameters:
      - description: Add Network Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_jail.AddJailNetworkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Add Network Switch to Jail
      tags:
      - Jail
  /jail/network/{ctId}/{networkId}:
    delete:
      consumes:
      - application/json
      description: Delete a network switch from a jail
      parameters:
      - description: Container ID
        in: path
        name: ctId
        required: true
        type: integer
      - description: Network ID
        in: path
        name: networkId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Delete Network Switch from Jail
      tags:
      - Jail
  /jail/network/inheritance/:ctId:
    put:
      consumes:
      - application/json
      description: Set network inheritance for a jail
      parameters:
      - description: Set Inheritance Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_jail.SetInheritanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Set Network Inheritance
      tags:
      - Jail
...
  /system/basic-settings:
    get:
      consumes:
      - application/json
      description: Get basic system settings information
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-github_com_alchemillahq_sylve_internal_db_models_BasicSettings'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Basic System Settings
      tags:
      - System
  /system/basic-settings/pools:
    put:
      consumes:
      - application/json
      description: Add usable ZFS pools to the system settings
      parameters:
      - description: List of ZFS Pools to add
        in: body
        name: pools
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Add Usable ZFS Pools
      tags:
      - System
  /system/basic-settings/toggle-service/:service:
    put:
      consumes:
      - application/json
      description: Enable or disable a specific service in the system settings
      parameters:
      - description: Service name to toggle
        in: path
        name: service
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Toggle Service
      tags:
      - System
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"